initial content shuffle
