+++ docs/advanced/Intro.js
@@ -1,0 +1,13 @@
+// @flow
+
+import React from 'react';
+import { H1 } from '../styled-components';
+
+export default function AdvancedIntro() {
+  return (
+    <div>
+      <H1>Advanced</H1>
+      <p>Content coming soon...</p>
+    </div>
+  );
+}
