+++ docs/advanced/index.js
@@ -1,0 +1,1 @@
+export { default as AdvancedIntro } from './Intro';
