+++ docs/async/Intro.js
@@ -1,0 +1,13 @@
+// @flow
+
+import React from 'react';
+import { H1 } from '../styled-components';
+
+export default function AsyncIntro() {
+  return (
+    <div>
+      <H1>Async</H1>
+      <p>Content coming soon...</p>
+    </div>
+  );
+}
