+++ docs/async/index.js
@@ -1,0 +1,1 @@
+export { default as AsyncIntro } from './Intro';
