+++ docs/creatable/Intro.js
@@ -1,0 +1,13 @@
+// @flow
+
+import React from 'react';
+import { H1 } from '../styled-components';
+
+export default function CreatableIntro() {
+  return (
+    <div>
+      <H1>Creatable</H1>
+      <p>Content coming soon...</p>
+    </div>
+  );
+}
