+++ docs/creatable/index.js
@@ -1,0 +1,1 @@
+export { default as CreatableIntro } from './Intro';
