+++ docs/guides/index.js
@@ -1,4 +1,2 @@
-export { default as Components } from './Components';
 export { default as Philosophy } from './Philosophy';
-export { default as Styles } from './Styles';
 export { default as QuickStart } from './QuickStart';
