+++ docs/home/Animated/Example.js
