+++ docs/home/Async/index.js
@@ -10,7 +10,7 @@ you also uncomment the <script> tag in ../index.html that loads the polyfill.
 */
 
 import React, { Component } from 'react';
-import { Code, H1 } from '../../components';
+import { Code, H1 } from '../../styled-components';
 import ExampleWrapper from '../../ExampleWrapper';
 import UsingCallbacks from './UsingCallbacks';
 import UsingPromises from './UsingPromises';
