+++ docs/home/Intro/Grouped.js
