+++ docs/home/Intro/index.js
@@ -1,14 +1,14 @@
 // @flow
 
 import React, { Component } from 'react';
-import { Code, H1, H2 } from '../../components';
+import { Code, H1, H2 } from '../../styled-components';
 
 import ExampleWrapper from '../../ExampleWrapper';
 import SingleSelect from './SingleSelect';
 import MultiSelect from './MultiSelect';
 import Grouped from './Grouped';
 
-export default class Home extends Component<*, *> {
+export default class HomeIntro extends Component<*, *> {
   render() {
     return (
       <div>
@@ -25,16 +25,16 @@ export default class Home extends Component<*, *> {
         </p>
         <ExampleWrapper
           label="Single Select"
-          urlPath="/docs/pages/Home/SingleSelect.js"
+          urlPath="/docs/home/Intro/SingleSelect.js"
         >
           <SingleSelect />
         </ExampleWrapper>
-        <ExampleWrapper label="Grouped" urlPath="/docs/pages/Home/Grouped.js">
+        <ExampleWrapper label="Grouped" urlPath="/docs/home/Intro/Grouped.js">
           <Grouped />
         </ExampleWrapper>
         <ExampleWrapper
           label="Multi Select"
-          urlPath="/docs/pages/Home/MultiSelect.js"
+          urlPath="/docs/home/Intro/MultiSelect.js"
         >
           <MultiSelect />
         </ExampleWrapper>
