+++ docs/home/Styled/SingleSelect.js
