+++ docs/home/index.js
@@ -2,5 +2,5 @@ export { default as Animated } from './Animated';
 export { default as Async } from './Async';
 export { default as Creatable } from './Creatable';
 export { default as Experimental } from './Experimental';
-export { default as Home } from './Home';
+export { default as Intro } from './Intro';
 export { default as Styled } from './Styled';
