+++ docs/styles/index.js
@@ -1,0 +1,1 @@
+export { default as StylesIntro } from './Intro';
